% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ThetasMmdl.R
\name{change_ThetasMmdl}
\alias{change_ThetasMmdl}
\title{Passes the metamodel to estimatesUI and, after returning, saves it as a metamodel file}
\usage{
change_ThetasMmdl(metamodelFile)
}
\arguments{
\item{metamodelFile}{File where the metamodel should be updated}
}
\value{
changed metamodel text
}
\description{
Passes the metamodel to estimatesUI and, after returning, saves it as a metamodel file
}
\examples{
if (interactive()) {
  OneCpt_IVInfusionFile <- system.file("vignettesdata/OneCpt_IVInfusion.mmdl",
  package = "Certara.RsNLME",
  mustWork = TRUE)
  
  change_ThetasMmdl(OneCpt_IVInfusionFile)
}

}
\keyword{internal}
