% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metamodelBuiltin.R
\name{create_metamodelBuiltin}
\alias{create_metamodelBuiltin}
\title{Run \code{modelBuilderUI()} and create resulting metamodel}
\usage{
create_metamodelBuiltin(metamodelFile, datafile, author = "")
}
\arguments{
\item{metamodelFile}{File where the metamodel should be created.}

\item{datafile}{File with input data.}

\item{author}{Optional character string to specify the author in the metamodel.}
}
\value{
NLME PML model S4 class instance
}
\description{
Used by Pirana to run \code{modelBuilderUI()}, saving the
resulting metamodel e.g., .mmmdl file given model building operations performed in GUI.
}
\examples{
if (interactive()) {
tmp_data <- tempfile(fileext = ".csv")
write.csv(Certara.RsNLME::pkData, tmp_data, row.names = FALSE)

create_metamodelBuiltin(
  "run1.mmdl",
  tmp_data
  )
}

}
