% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{NlmeCovarItem-class}
\alias{NlmeCovarItem-class}
\alias{NlmeCovarItem}
\title{Class represents an NLME Covariate parameter}
\arguments{
\item{name}{Name of covariate}

\item{value}{Covariate value}
}
\description{
Class represents an NLME Covariate parameter
}
\keyword{internal}
